Add tests for Kleene star.
Two things: test ** and test empty matches at the start of strings.

This surprised me, but perhaps it shouldn't have.
The expression a* matches the empty string at the beginning of "xyzaaa."
I expected it to match the a's.
However, that's not how Acme behaves.
In retrospect, it's better that it matches the empty,
otherwise the behavior is inconsistent
(why doesn't * return its earliest match?)
and it would be difficult to describe.