Make the ui.Canvas API stateless.
This removes the SetColor method.
There are only two methods that need the color;
they can just take it as an argument.

I also swapped the order of the arguments to Canvas.Draw.
Fill and Stroke take the color first and the location second.
Draw should take the image, which specifies the colors,
first and the location second too.