Hold a copy of the texture data to initialize the pixels when locking.
SDL_LockTexture returns uninitialied memory.
It doesn't even contain the previous texture data.
So, we need to hold on to our own copy
to restore it when the texture is locked.
This is what the SDL wiki recommends.

This means that the first Set called on
a texture that was previously drawn
will be slow as it will blit the image data
to the texture memory.
The alternative is to comment that reading
from the texture image is undefined.
It's better to have slower-but-defined behavior.

Fixes #62.