Change the runes.Read method to return the runes.
All of the callers, with the exception of a benchmark, allocated the rune slice right before calling read. Read might as well just return the slice. Then it can handle out-of-bounds reads the same way as Insert and Delete: it can error. All of the callers ensure that they are reading in bounds before calling it anway.