Add runes.Copy.
The fast path directly uses the destination Writer's cache as the Reader's buffer, eliminating an extra copy. This case fires if we can determine the exact number of bytes to be read. We can determine this if the reader is a buffer.Reader(·), SliceReader, or a LimitedReader for which we can determine the size of the underlying reader. This should cover the common cases.