Change Editor.do not to be transactional.
Before this change, Editor.do performed retrying transactions.
It worked in three steps:
1) Take the read lock and compute the change into ed.pending.
2) Release the read lock and fix the change addresses to be locally consistent.
3) Take the write lock.
	• If the sequence number changed since #1, release the lock, drop the changes and retry at #1
	• Otherwise, apply the changes in ed.pending and release the lock.

This allowed multiple readers to make changes in parallel, but required that all changes computed in #1 be idempotent, as #1 may be performed multiple times in the face of retries. Pipe edits cannot be idempotent. They must read from or write to external programs, and they can only do it once—no retries.

This change restructures Editor.do to remove the transactions. Now, it simply takes a lock, computes the change, fixes the addresses, applies the change, and releases the lock.

I also added some tests to ensure that marks are correctly restored in the face of errors.