Redefine the Set Edit to eliminate the possibility of errors.
Set used to impose restrictions on which runes it allowed. There was no practical reason for this. All it did was introduce the possibility of errors: it's an error for the user to specify an invalid rune.

No Edit creating functions return errors. They all return one thing: an Edit. It is not worth breaking this for Set, so I redefined it to allow any rune.

The parsing done by Ed skips whitespace between k and the mark name rune. If there is only whitespace following k, then it is interpreted as no mark name, and uses the default of dot. To maintain consistency with that, whitespace marks passed to Set are interpreted as dot.