Fix escape handling in edits.
Change and Substitute edits both need to interpret an escaped string of text.
For Change edits, it's the change-to string.
For Substitute edits, it's the replace-with string.
Substitute edits never properly escaped an added delimiter in the replace-with string in the String method.
Beyond that, the escape handling for both was written separately and handled hap-hazardly.

This change unifies the escape handling and fixes the unescaped delimiter bug in Substitute edits.

Fixes #156.