Port edit tests to a new test framework.
The new framework uses a simple notation to describe buffers and their marks. This has two benefits: 1) it's easier to read than numbers like addr{5, 62}, and 2) it doesn't rely on internal details, for example, that addr{} currently holds rune offsets.

The new framework also allows for multiple Edits, so there is no need for a separate Undo/Redo test framework.

The old Address test framework was testing the internal interface by calling the whereFrom method of Address. This was masking various bugs (#178, #179, #180, and #182). This change re-implements the Address tests using the new editTest framework. This way, there is only a single test framework for testing Edits and Addresses, and the Address tests test the external interface.

Fixes #173.