Evaluate regexp addresses relative to . and don't compute ; as +.
Regexp addresses were only being evaluated relative to 0 or $. But they should be relative to . instead. Also ; would set dot to a1 and evaluate +a2 for the end of the range. Instead, set dot to a1 and evaluate a2 relative to its default (which is most-often `0`, but is `ed.marks['.'].to` for regexps).

Fixes #180.
Fixes #179.