re1: don't return negative-sized submatches.
This was happening in the case of nested subexpressions.
If the outter subexpression matched multiple times, but the inner expression did not match on the final match of the outter, then the inner expression's submatch would have a negative size: its end was set on the earlier match, but it's start was set on the later match. Now, whenever setting the start, reset the end of the match too.

Practically speaking this probably doesn't matter, but it can simplify calling code to know that submatches are always non-negative.