edit: defer reporting Editor.Change errors until Editor.Apply.
This cuts down on the error boilerplate code in the Edit implementations.
I also added setDot, which sets the '.' mark and panics on error.
It is called by Edit implementations after Address.Where computes a Span.
Such Spans can't be out of range or else Address.Where would have returned the error.