edit: allow empty regexp address matches.
Previously they were "no match" errors.
The behavior differs from acme and sam, both of which always return the second match in the case of empty matches. For example `/b*/` on `{..}aaa` would return the empty string after the first a. Maybe that behavior is convenient for something, but I'm not sure what. T always returns the first empty match: the empty string before the first a.

Fixes #217.