ui/text: don't double-fill.
Previously, we filled in the background of a text area, then re-filled and drew the lines over it. Instead, just fill and draw the lines and then fill the borders separately. This means that we fill less. It should also reduce the flashing on x11, which currently isn't truly double-buffered.