ui: be more strict about the minimum redraw time.
The previous version used a ticker channel. The ticker restarts right after sending to the channel. This meant that drawing a particularly slow frame could fill the channel and redraws to happen much faster than the maximum rate. Instead, use a timer that is reset *after* we finish drawing the current frame. This enforces the minimum time between the *end* of one frame and the *start* of the next.