ui/text: fix bug with tabs + padding.
Tabs go to the next tab stop.
There was an inconsistency between how drawLines computed tab stops
and how both Text.GlyphBox and Text.Index computed tab stops.
drawLines did not include the padding in the tabstop measure; the Text methods did.
The correct behavior is that of drawLines.

Fixes #279.