editor/view,ui: return errors preforming edits on a View.
We were returning errors from individual edits, but if the entire HTTP request failed, no error was returned, and the response channel was never notified. This caused any such error to deadlock waiting for a response that would never be sent. Now, the interface is changed to hide the underlying channel and simply return the response or an error instead.

This change also makes the ui code more explicit about whether it's doing a synchronous or asynchornous request. This is important, because we want to avoid blocking the window's event goroutine on synchronous requests, which we currently do in a few places (see #295).

Fixes #293.